+++ app/controllers/pdl.py
@@ -15,6 +15,8 @@ from app.models.state import State
 from app.models.statecode import StateCode
 from app.models.subprogram import Subprogram
 from app.models.subtitle import Subtitle
+from app.models.title import Title
+from collections import defaultdict
 
 LANDING_PAGE_DATA_PATH = os.path.join("controllers", "data", "landingpage")
 ALLPROGRAM_DATA_JSON = "allprograms.json"
@@ -34,9 +36,14 @@ SADA_SUBPROGRAMS_DATA_JSON = "sada_subprograms_data.json"
 
 TITLE_II_DATA_PATH = os.path.join("controllers", "data", "title-ii")
 II_EQIP_DATA_PATH = os.path.join(TITLE_II_DATA_PATH, "programs", "eqip")
+II_EQIP_IRA_DATA_PATH = os.path.join(TITLE_II_DATA_PATH, "programs", "eqip-ira")
 EQIP_MAP_DATA_JSON = "eqip_map_data.json"
 EQIP_STATE_DISTRIBUTION_DATA_JSON = "eqip_state_distribution_data.json"
 EQIP_PRACTICE_CATEGORIES_DATA_JSON = "eqip_practice_categories_data.json"
+EQIP_IRA_STATE_DISTRIBUTION_DATA_JSON = "eqip_ira_state_distribution.json"
+EQIP_IRA_SUMMARY_DATA_JSON = "eqip_ira_summary.json"
+EQIP_IRA_PRACTICE_CATEGORIES_DATA_JSON = "eqip_ira_practices.json"
+EQIP_IRA_AGGREGATE_DATA_JSON = "eqip_ira_aggregated_prediction.json"
 II_CSP_DATA_PATH = os.path.join(TITLE_II_DATA_PATH, "programs", "csp")
 CSP_MAP_DATA_JSON = "csp_map_data.json"
 CSP_STATE_DISTRIBUTION_DATA_JSON = "csp_state_distribution_data.json"
@@ -190,6 +197,24 @@ def allprograms_search():
         return data_json
 
 
+# /pdl/titles/title-i/summary:
+def titles_title_i_summary_search():
+    start_year = 2014
+    end_year = 2021
+    title_id = 100
+    endpoint_response = generate_title_i_total_summary_response(title_id, start_year, end_year)
+    return endpoint_response
+
+
+# /pdl/titles/title-i/state-distribution:
+def titles_title_i_state_distribution_search():
+    start_year = 2014
+    end_year = 2021
+    title_id = 100
+    endpoint_response = generate_title_i_total_state_distribution_response(title_id, start_year, end_year)
+    return endpoint_response
+
+
 # /pdl/titles/title-i/subtitles/subtitle-a/map:
 def titles_title_i_subtitles_subtitle_a_map_search():
     # set the file path
@@ -296,6 +321,66 @@ def titles_title_ii_programs_eqip_summary_search():
 
     # open file
     with open(eqip_data, 'r') as map_data:
+        file_data = map_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+
+# /pdl/titles/title-ii/programs/eqip-ira/state-distribution
+def titles_title_ii_programs_eqip_ira_state_distribution_search():
+    # set the file path
+    eqip_ira_data = os.path.join(II_EQIP_IRA_DATA_PATH, EQIP_IRA_STATE_DISTRIBUTION_DATA_JSON)
+
+    # open file
+    with open(eqip_ira_data, 'r') as map_data:
+        file_data = map_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+
+# /pdl/titles/title-ii/programs/eqip-ira/summary
+def titles_title_ii_programs_eqip_ira_summary_search():
+    # set the file path
+    eqip_ira_data = os.path.join(II_EQIP_IRA_DATA_PATH, EQIP_IRA_SUMMARY_DATA_JSON)
+
+    # open file
+    with open(eqip_ira_data, 'r') as map_data:
+        file_data = map_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+
+# /pdl/titles/title-ii/programs/eqip-ira/practice-names
+def titles_title_ii_programs_eqip_ira_practice_names_search():
+    # set the file path
+    eqip_ira_data = os.path.join(II_EQIP_IRA_DATA_PATH, EQIP_IRA_PRACTICE_CATEGORIES_DATA_JSON)
+
+    # open file
+    with open(eqip_ira_data, 'r') as map_data:
+        file_data = map_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+
+# /pdl/titles/title-ii/programs/eqip-ira/predicted
+def titles_title_ii_programs_eqip_ira_predicted_search():
+    # set the file path
+    eqip_ira_data = os.path.join(II_EQIP_IRA_DATA_PATH, EQIP_IRA_AGGREGATE_DATA_JSON)
+
+    # open file
+    with open(eqip_ira_data, 'r') as map_data:
         file_data = map_data.read()
 
         # parse file
@@ -471,32 +556,11 @@ def titles_title_xi_programs_crop_insurance_summary_search():
 
 # /pdl/titles/title-iv/programs/snap/state-distribution
 def titles_title_iv_programs_snap_state_distribution_search():
-    # set the file path
-    snap_data = os.path.join(IV_SNAP_DATA_PATH, SNAP_STATE_DISTRIBUTION_DATA_JSON)
-
-    # open file
-    with open(snap_data, 'r') as map_data:
-        file_data = map_data.read()
-
-        # parse file
-        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
-
-        return data_json
-
-
-# /pdl/titles/title-iv/programs/snap/summary
-def titles_title_iv_programs_snap_summary_search():
-    # set the file path
-    snap_data = os.path.join(IV_SNAP_DATA_PATH, SNAP_SUBPROGRAMS_DATA_JSON)
-
-    # open file
-    with open(snap_data, 'r') as map_data:
-        file_data = map_data.read()
-
-        # parse file
-        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
-
-        return data_json
+    program_id = 106
+    start_year = 2018
+    end_year = 2022
+    endpoint_response = generate_title_iv_state_distribution_response(program_id, start_year, end_year)
+    return endpoint_response
 
 
 # construct state
@@ -568,6 +632,256 @@ def construct_allprogram_result(prog):
     }
 
     return result
+
+
+def generate_title_iv_state_distribution_response(program_id, start_year, end_year):
+    session = Session()
+
+    # construct the query
+    program_query = session.query(
+        Payment.state_code.label('state'),
+        Program.name.label('programName'),
+        Payment.year.label('year'),
+        Payment.payment.label('totalPaymentInDollars'),
+        Payment.recipient_count.label('totalCounts')).join(
+        Program, Payment.program_id == Program.id).filter(
+        Payment.program_id == program_id,
+        Payment.year.between(start_year, end_year)
+    )
+
+    # execute the query
+    program_result = program_query.all()
+
+    # create dictionaries
+    program_response_dict = defaultdict(list)
+    year_dict = defaultdict(list)
+    total_national_payment_dict = defaultdict(float)
+    total_national_count_dict = defaultdict(int)
+
+    # aggregate data
+    for record in program_result:
+        year = str(record[2])
+        state = record[0]
+        payment = record[3]
+        recipient_count = record[4]
+
+        year_dict[year].append({
+            'state': state,
+            'totalPaymentInDollars': payment,
+            'averageMonthlyParticipation': recipient_count
+        })
+
+        # update total national payment and count for the year
+        total_national_payment_dict[year] += payment
+        total_national_count_dict[year] += recipient_count
+
+    # create output response dictionary
+    for year, data in year_dict.items():
+        program_response_dict[year] = []
+        total_payment_nationwide = total_national_payment_dict[year]
+        total_participation_nationwide = total_national_count_dict[year]
+        for item in data:
+            state = item['state']
+            total_payment = item['totalPaymentInDollars']
+            average_participation = item['averageMonthlyParticipation']
+
+            total_payment_percentage = (total_payment / total_payment_nationwide) * 100
+            average_participation_percentage = (average_participation / total_participation_nationwide) * 100
+
+            program_response_dict[year].append({
+                'state': state,
+                'totalPaymentInDollars': total_payment,
+                'totalPaymentInPercentageNationwide': round(total_payment_percentage, 2),
+                'averageMonthlyParticipation': average_participation,
+                'averageMonthlyParticipationInPercentageNationwide': round(average_participation_percentage, 2)
+            })
+
+        # sort states by decreasing order of total payment in dollars
+        program_response_dict[year] = \
+            sorted(program_response_dict[year], key=lambda x: x['totalPaymentInDollars'], reverse=True)
+
+    # calculate total payment and average monthly participation for each state across all years
+    state_total_data = defaultdict(lambda: {'totalPaymentInDollars': 0, 'averageMonthlyParticipation': 0})
+
+    for year, data in year_dict.items():
+        for item in data:
+            state = item['state']
+            total_payment = item['totalPaymentInDollars']
+            average_participation = item['averageMonthlyParticipation']
+
+            state_total_data[state]['totalPaymentInDollars'] += total_payment
+            state_total_data[state]['averageMonthlyParticipation'] += average_participation
+
+    # calculate the sum of total national payment and count for all years
+    total_national_payment = sum(total_national_payment_dict.values())
+    total_national_count = sum(total_national_count_dict.values())
+
+    # add the total data for each state to the program_response_dict
+    total_states_data = []
+    for state, total_data in state_total_data.items():
+        total_payment = total_data['totalPaymentInDollars']
+        total_participation = total_data['averageMonthlyParticipation']
+
+        #  calculate percentage using the total national payment and count across all years
+        total_payment_percentage = (total_payment / total_national_payment) * 100
+        total_participation_percentage = (total_participation / total_national_count) * 100
+
+        # append total data for each state to the list
+        total_states_data.append({
+            'state': state,
+            'totalPaymentInDollars': total_payment,
+            'totalPaymentInPercentageNationwide': round(total_payment_percentage, 2),
+            'averageMonthlyParticipation': total_participation,
+            'averageMonthlyParticipationInPercentageNationwide': round(total_participation_percentage, 2)
+        })
+
+    # Sort the total states data by decreasing order of total payment in dollars
+    total_states_data = sorted(total_states_data, key=lambda x: x['totalPaymentInDollars'], reverse=True)
+
+    # Add the total states data to the program_response_dict under the key 'TotalStates'
+    program_response_dict[str(start_year) + '-' + str(end_year)] = total_states_data
+
+    return program_response_dict
+
+
+def generate_title_i_total_state_distribution_response(title_id, start_year, end_year):
+    session = Session()
+
+    # construct the query
+    program_query = session.query(
+        Payment.state_code.label('state'),
+        Subtitle.name.label('subtitleName'),
+        Payment.year.label('year'),
+        Payment.payment.label('totalPaymentInDollars'),
+        Payment.recipient_count.label('totalCounts')).join(
+        Subtitle, Payment.subtitle_id == Subtitle.id).filter(
+        Payment.title_id == title_id,
+        Payment.year.between(start_year, end_year)
+    )
+
+    # execute the query
+    result = program_query.all()
+
+    # create a nested dictionary to store data by year and state
+    data_by_year_and_state = defaultdict(
+        lambda: defaultdict(lambda: {'totalPaymentInDollars': 0, 'totalRecipients': 0}))
+    all_years_summary = defaultdict(lambda: {'totalPaymentInDollars': 0, 'totalRecipients': 0})
+
+    for record in result:
+        state, title_name, year, payments, recipients = record
+        entry = data_by_year_and_state[year][state]
+        entry['state'] = state
+        entry['totalPaymentInDollars'] += payments
+        entry['totalRecipients'] += recipients
+
+        # add to all years summary
+        summary = all_years_summary[state]
+        summary['state'] = state
+        summary['totalPaymentInDollars'] += payments
+        summary['totalRecipients'] += recipients
+
+    # sort by total payment
+    sorted_data_by_year = {}
+    for year, states in data_by_year_and_state.items():
+        sorted_entries = sorted(states.values(), key=lambda x: x['totalPaymentInDollars'], reverse=True)
+        sorted_data_by_year[year] = sorted_entries
+
+    # round the total payment
+    for year, states in sorted_data_by_year.items():
+        for state in states:
+            state['totalPaymentInDollars'] = round(state['totalPaymentInDollars'], 2)
+
+    # all years summary
+    sorted_summary = sorted(all_years_summary.values(), key=lambda x: x['totalPaymentInDollars'], reverse=True)
+
+    # round the total payment
+    for state in sorted_summary:
+        state['totalPaymentInDollars'] = round(state['totalPaymentInDollars'], 2)
+
+    sorted_data_by_year['2014-2021'] = sorted_summary
+
+    result_dict = dict(sorted_data_by_year)
+
+    return result_dict
+
+
+def generate_title_i_total_summary_response(title_id, start_year, end_year):
+    session = Session()
+
+    # construct the query
+    program_query = session.query(
+        Payment.state_code.label('state'),
+        Title.name.label('titleName'),
+        Payment.year.label('year'),
+        Payment.payment.label('totalPaymentInDollars'),
+        Payment.recipient_count.label('totalCounts'),
+        Subtitle.name.label('subtitleName')
+    ).join(
+        Title, Payment.title_id == Title.id
+    ).join(
+        Subtitle, Payment.subtitle_id == Subtitle.id
+    ).filter(
+        Payment.title_id == title_id,
+        Payment.year.between(start_year, end_year)
+    )
+
+    # execute the query
+    results = program_query.all()
+
+    # Initialize dictionaries to store aggregated data
+    title_summary = defaultdict(lambda: {
+        'totalPaymentInDollars': 0,
+        'totalCounts': 0,
+        'recipients': [],
+        'subtitles': defaultdict(lambda: {
+            'totalPaymentInDollars': 0,
+            'totalCounts': 0,
+            'recipients': []
+        })
+    })
+
+    # Process each record in the data
+    for state, title_name, year, payment, recipients, subtitle in results:
+        title_summary[title_name]['totalPaymentInDollars'] += payment
+        title_summary[title_name]['totalCounts'] += recipients
+        title_summary[title_name]['recipients'].append(recipients)
+
+        # Aggregate subtitle data
+        subtitle_dict = title_summary[title_name]['subtitles'][subtitle]
+        subtitle_dict['totalPaymentInDollars'] += payment
+        subtitle_dict['totalCounts'] += recipients
+        subtitle_dict['recipients'].append(recipients)
+
+    # Prepare the final summary
+    final_summary = []
+    for title, info in title_summary.items():
+        average_recipient_count = sum(info['recipients']) / len(info['recipients']) if info['recipients'] else 0
+        subtitle_list = []
+        total_payment_title = info['totalPaymentInDollars']
+
+        for subtitle_name, subtitle_info in info['subtitles'].items():
+            subtitle_avg_recipients = sum(subtitle_info['recipients']) / len(subtitle_info['recipients']) if \
+            subtitle_info['recipients'] else 0
+            payment_percentage = (subtitle_info[
+                                      'totalPaymentInDollars'] / total_payment_title * 100) if total_payment_title else 0
+            subtitle_list.append({
+                'programName': subtitle_name,
+                'totalPaymentInDollars': round(subtitle_info['totalPaymentInDollars'], 2),
+                'totalCounts': subtitle_info['totalCounts'],
+                'averageRecipientCount': subtitle_avg_recipients,
+                'totalPaymentInPercentage': round(payment_percentage, 2)
+            })
+
+        title_entry = {
+            'titleName': title,
+            'totalPaymentInDollars': round(info['totalPaymentInDollars'], 2),
+            'totalCounts': info['totalCounts'],
+            'averageRecipientCount': round(average_recipient_count, 2),
+            'subtitles': subtitle_list
+        }
+        final_summary.append(title_entry)
+
+    return final_summary
 
 
 def generate_title_i_state_distribution_response(subtitle_id, start_year, end_year):